requirements.txt
requests
django
gunicorn
django-heroku
selenium
djangorestframework